+++ tests/providers/test_factory.py
@@ -2,7 +2,7 @@ import pytest
 
 from antidote.core import DependencyContainer
 from antidote.exceptions import DuplicateDependencyError
-from antidote.providers.service import Build, ServiceProvider, LazyFactory
+from antidote.providers.factory import Build, FactoryProvider
 
 
 class Service:
@@ -19,7 +19,7 @@ class AnotherService:
 @pytest.fixture()
 def provider():
     container = DependencyContainer()
-    provider = ServiceProvider(container=container)
+    provider = FactoryProvider(container=container)
     container.register_provider(provider)
     return provider
 
@@ -61,59 +61,58 @@ def test_invalid_build(args: tuple, kwargs: dict):
         Build(*args, **kwargs)
 
 
-def test_simple(provider: ServiceProvider):
-    provider.register(factory=Service, service=Service)
+def test_simple(provider: FactoryProvider):
+    provider.register_class(Service)
 
     dependency = provider.provide(Service)
     assert isinstance(dependency.instance, Service)
     assert repr(Service) in repr(provider)
 
 
-def test_singleton(provider: ServiceProvider):
-    provider.register(factory=Service, service=Service, singleton=True)
-    provider.register(factory=AnotherService, service=AnotherService,
-                      singleton=False)
+def test_singleton(provider: FactoryProvider):
+    provider.register_class(Service, singleton=True)
+    provider.register_class(AnotherService, singleton=False)
 
     provide = provider.provide
     assert provide(Service).singleton is True
     assert provide(AnotherService).singleton is False
 
 
-def test_takes_dependency(provider: ServiceProvider):
-    provider.register(factory=lambda cls: cls(), service=Service,
-                      takes_dependency=True)
+def test_takes_dependency(provider: FactoryProvider):
+    provider.register_factory(factory=lambda cls: cls(), dependency=Service,
+                              takes_dependency=True)
 
     assert isinstance(provider.provide(Service).instance, Service)
     assert provider.provide(AnotherService) is None
 
 
-def test_build(provider: ServiceProvider):
-    provider.register(factory=Service, service=Service)
+def test_build(provider: FactoryProvider):
+    provider.register_class(Service)
 
     s = provider.provide(Build(Service, 1, val=object)).instance
     assert isinstance(s, Service)
     assert (1,) == s.args
     assert dict(val=object) == s.kwargs
+#
+#
+# def test_lazy(provider: FactoryProvider):
+#     sentinel = object()
+#     provider._container.update_singletons({
+#         'lazy_getter': lambda: sentinel
+#     })
+#     provider.register_factory(factory=LazyFactory('lazy_getter'), dependency=Service)
+#
+#     assert sentinel is provider.provide(Service).instance
 
 
-def test_lazy(provider: ServiceProvider):
-    sentinel = object()
-    provider._container.update_singletons({
-        'lazy_getter': lambda: sentinel
-    })
-    provider.register(factory=LazyFactory('lazy_getter'), service=Service)
-
-    assert sentinel is provider.provide(Service).instance
-
-
-def test_duplicate_error(provider: ServiceProvider):
-    provider.register(factory=Service, service=Service)
-
-    with pytest.raises(DuplicateDependencyError):
-        provider.register(factory=Service, service=Service)
+def test_duplicate_error(provider: FactoryProvider):
+    provider.register_class(Service)
 
     with pytest.raises(DuplicateDependencyError):
-        provider.register(factory=lambda: Service(), service=Service)
+        provider.register_class(Service)
+
+    with pytest.raises(DuplicateDependencyError):
+        provider.register_factory(factory=lambda: Service(), dependency=Service)
 
 
 @pytest.mark.parametrize(
@@ -123,11 +122,11 @@ def test_duplicate_error(provider: ServiceProvider):
      dict(factory='test', service=Service),
      dict(factory=object(), service=Service)]
 )
-def test_invalid_type(provider: ServiceProvider, kwargs):
+def test_invalid_type(provider: FactoryProvider, kwargs):
     with pytest.raises(TypeError):
-        provider.register(**kwargs)
+        provider.register_factory(**kwargs)
 
 
 @pytest.mark.parametrize('dependency', ['test', Service, object()])
-def test_unknown_dependency(provider: ServiceProvider, dependency):
+def test_unknown_dependency(provider: FactoryProvider, dependency):
     assert provider.provide(dependency) is None